refactor(ui): merge ACME servers view into TLS certificates tabs
Replace the top-right navigation button with two tabs, Certificates and
ACME settings, on the TLS certificates page. The ACME servers page becomes
a child component of that page and its route redirects to the new tab.

Reorder the settings table columns so that the node comes first, and rename
the section to "ACME settings".

The certificates table is only re-indented, its content is unchanged.
